backend/pyproject.toml
[tool.poetry]
name = "aiogram-soloway-bot"
version = "0.1.0"
description = ""
authors = ["Vlad Laptev"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "4.0"
python-dotenv = "0.17.1"
psycopg2-binary = "2.9.3"
asyncpg = "0.23.0"
djangorestframework = "3.13.1"
requests = "2.27.1"
redis = "4.3.3"
celery = "5.2.3"
gunicorn = "20.1.0"
python-telegram-bot = "13.12"
drf-yasg = "1.20.0"
httpx = "0.23.0"
asgiref = "3.5.2"
pydantic = "1.9.2"
backoff = "2.1.2"
django-environ = "0.9.0"
django-celery-beat = "^2.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "loader.TORTOISE_CONFIG"
location = "./db_models/migrations"
src_folder = "./."